Move StoreTrainingData to StoreTrainingData.py
